# Reduce decoding allocations