@DavidVorick 2017-08-31
I think I'd prefer a complete solution via a fully custom `UnmarshalSia` for `types.Block`. Can you write a benchmark that demonstrates a performance improvement?