@lukechampine 2017-09-01
I benchmarked by creating and unlocking a wallet from scratch (rescanning ~120,000 blocks).

Naive decoder: 6m2s, 58.8 GB GB allocated
Custom decoder: 4m44s, 48.6 GB allocated

i.e. ~20% faster unlock speed and ~20% less memory allocated.

Profiling indicates that most of the remaining allocations are "necessary," e.g. allocating the various slices in the `types.Transaction` struct. We could eliminate these in theory (consensus could reuse the memory of previously-decoded blocks) but it would be ugly and questionably safe.

I used a somewhat novel strategy to minimize allocations: threading a `decHelper` struct through methods that take an `io.Reader`. Each `UnmarshalSia` method calls `decoder` on the `io.Reader`; if the underlying type is a `decHelper`, it returns it, otherwise it creates and returns a new `decHelper`. This also allows me to use helper methods within the various `UnmarshalSia` methods. The error handling approach is monadic, accumulating errors internally and returning them at the end with `d.Err()`. Probably this should be reworked to store the first non-nil error (instead of the most recent error), and to return immediately from all methods if the stored error is non-nil.

Finally, this code is a bit less safe than the reflection-based decoder in one regard: it doesn't protect against unacceptably large slice lengths. That is, if an encoded object tells you that it's a slice with 2^64 elements, this new code will go right ahead and try to allocate it, whereas the reflection-based decoder will not. I plan to restore this bit of sanity checking, but it will take a bit of work.