@rudibs 2017-09-01
i wonder if that unacceptably large slice length guard difference represents a network split attack vector for e.g. double spending