@DavidVorick 2017-09-01
I was looking for a golang benchmark lol.

Though, shaving more than 60 seconds off of the full unlock time is unquestionably very impressive and worth merging. We should still have a golang benchmark for block encoding though. We do have one for empty blocks, though they don't seem to have been affected by the full 20% that you are reporting for natural blocks (which is perhaps not surprising).