@DavidVorick 2017-09-01
@rudibs if you change the longest allowed slice, then yes that introduces a hardfork. In this case though, it's only being added to the string field, which as best I can remember there are no strings in the consensus types, so we should be okay.