@DavidVorick 2017-09-01
Overall this looks good, but just like with the other MarshalSia implementations we wrote, we need a sanity check in here that verifies the marshalling pattern matches the old code exactly - this is consensus code and mistakes will cause hardforks.