@lukechampine 2017-09-01
It just occurred to me that I can use the `decoder` trick to reduce allocations in `MarshalSia` too. Right now one of the largest sources of allocations is `encoding.WriteUint64`, because it has to allocate a new slice for each uint64 it writes. But if we're threading an `encoder` type through the `MarshalSia` methods, we can reuse a single buffer to encode all the writes. I'll do that in a follow up PR.