@lukechampine 2017-09-01
Benchmarks:
```
old:

BenchmarkDecodeEmptyBlock-4	 2000000    750 ns/op	 85.26 MB/s    6 allocs/op
BenchmarkDecodeHeavyBlock-4	  100000  19735 ns/op 	 72.87 MB/s  167 allocs/op

new:

BenchmarkDecodeEmptyBlock-4	 5000000    384 ns/op	166.45 MB/s    3 allocs/op
BenchmarkDecodeHeavyBlock-4	  300000   4398 ns/op	326.93 MB/s   51 allocs/op
```

The new test adds coverage for all the new `UnmarshalSia` methods.