@rudibs 2017-09-01
Great logic David and Luke!!  You really explained it well thanks.