@lukechampine 2017-09-01
ok, maximum sizes are now enforced. This should be ready for final review